Adding RTOS to pico2 with trustzone

+++ compile.sh
@@ -1,28 +1,46 @@
 #!/bin/bash
 set -e  # Exit immediately on any command failure
 
-rm -rf build_secure
-rm -rf build_nonsecure
-mkdir build_secure
-mkdir build_nonsecure
+CMAKECMD=cmake
+GCC=~/.pico-sdk/toolchain/14_2_Rel1/bin/arm-none-eabi-gcc
+OBJDUMP=~/.pico-sdk/toolchain/14_2_Rel1/bin/arm-none-eabi-objdump 
+
+# if argument is -clean
+if [ "$1" == "--clean" ]; then
+    echo "[LOG] Cleaning build directories..."
+    rm -rf build_secure
+    rm -rf build_nonsecure
+fi
+
+
+# if build_secure dont exist
+if [ ! -d "build_secure" ]; then
+    mkdir build_secure
+fi
+if [ ! -d "build_nonsecure" ]; then
+    mkdir build_nonsecure
+fi
 
 # # --- Secure Build ---
+echo "[LOG] Building secure firmware..."
 cd build_secure
-cmake -G Ninja .. -DCMAKE_TOOLCHAIN_FILE=../toolchain/rp2350.cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DBUILD_SECURE=1
-cmake --build . --verbose #| tee ../secure_build.log
-# if return code is 0, then run the next command else exit
-
+$CMAKECMD -G Ninja .. -DCMAKE_TOOLCHAIN_FILE=../toolchain/rp2350.cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DBUILD_SECURE=1
+$CMAKECMD --build . --verbose #| tee ../secure_build.log
 cd ..
+echo "[LOG] Secure build completed."
 
 # --- Non-Secure Build ---
+echo "[LOG] Building non-secure firmware..."
 cd build_nonsecure
-cmake -G Ninja .. -DCMAKE_TOOLCHAIN_FILE=../toolchain/rp2350.cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DBUILD_NONSECURE=1
-cmake --build . --verbose #| tee ../nonsecure_build.log
+$CMAKECMD -G Ninja .. -DCMAKE_TOOLCHAIN_FILE=../toolchain/rp2350.cmake -DCMAKE_EXPORT_COMPILE_COMMANDS=ON -DBUILD_NONSECURE=1
+$CMAKECMD --build . --verbose #| tee ../nonsecure_build.log
 cd ..
-
-
+echo "[LOG] Nonsecure build completed."
 
 # --- Objdump (only after both builds are done) ---
-echo "Generating disassembly files..."
-arm-none-eabi-objdump -D build_secure/secure/secure.elf > build_secure/secure/secure.elf.dis
-arm-none-eabi-objdump -D build_nonsecure/nonsecure/nonsecure.elf > build_nonsecure/nonsecure/nonsecure.elf.dis
+echo "[LOG] Generating disassembly files..."
+$OBJDUMP -D build_secure/secure/secure.elf > build_secure/secure/secure.elf.dis
+$OBJDUMP -D build_nonsecure/nonsecure/nonsecure.elf > build_nonsecure/nonsecure/nonsecure.elf.dis
+echo "[LOG] Disassembly files generated."
+
+echo "[LOG] Build process completed."
